refactor: refactor auth module directory structure
