+++ server/src/auth/auth.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Get, Post, Request, UseGuards } from "@nestjs/common";
 
 import { AuthService } from "./auth.service";
-import { JwtAuthGuard } from "./jwt.guard";
-import { LocalAuthGuard } from "./local.guard";
+import { JwtAuthGuard } from "./guards/jwt.guard";
+import { LocalAuthGuard } from "./guards/local.guard";
 
 @Controller("auth")
 export class AuthController {
